backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
python-multipart>=0.0.12
psycopg2-binary>=2.9.9
openai>=1.0.0
python-dotenv>=1.0.0
sse-starlette>=2.0.0
